Patch from jbeuree (modified to work better) to fix seeking on large files with the ivtv decoder.
Fixes #68.


git-svn-id: http://svn.mythtv.org/svn/trunk@6786 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/ivtvdecoder.cpp
@@ -535,11 +535,11 @@ void IvtvDecoder::UpdateFramesPlayed(void)
     m_parent->SetFramesPlayed(videoPlayed);
 }
 
-bool IvtvDecoder::StepFrames(int start, int count)
+bool IvtvDecoder::StepFrames(long long start, long long count)
 {
     VideoOutputIvtv *videoout = (VideoOutputIvtv *)m_parent->getVideoOutput();
 
-    int step, cur = 0, last = start;
+    long long step, cur = 0, last = start;
 
     //fprintf(stderr, "stepping %d from %d\n", count, last);
